pyproject.toml
[tool.poetry]
name = "pysinker"
version = "0.1.0"
description = "A simple tool automate database backup"
authors = ["Orion WAMBERT <wambert.orion@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
typer = "^0.12.5"
setuptools = "^73.0.1"
pyyaml = "^6.0.2"
schedule = "^1.2.2"
pytest-watch = "^4.2.0"


apscheduler = "^3.10.4"
python-daemon = "^3.0.1"
boto3 = "^1.35.22"
[tool.poetry.group.dev.dependencies]
py-mon = "^2.0.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
pysinker = "pysinker.main:main"
